New data type: JSONClass; additions to JSONable.
JSONClass is a class (as opposed to JSONable, which is an interface). As
such, JSONClass can do more than JSONable. It can define a constructor,
JSONClass(JSONObject json) for json-loading, and an actually implemented
method, JSONObject toJSON() for json-saving. The loading and saving is
incomplete at the moment. It only saves and loads primitive types and
Strings. It also only saves and loads variables defined within the child
class itself, so any variables that are defined in the child's parent
will not be automatically saved.

Some new default methods were added to JSONable to fill in some gaps.